filters by year already work (but without navigation)
Use `?year=2025` on the url to filter. In the future, a navigation bar
will allow visitors to choose the year themselves, without needing to
fiddle with the URL.

Also, in the future a message mentioning this script will show up on
the footer, but you'll be able to hide it if you prefer not to have
that mentioned.